refactor(toast): migrate to Ark v1 toast API
