+++ src/lib/types/storybook/Tags.enum.ts
@@ -6,8 +6,6 @@
 enum Tags {
   /** Enable [autodocs](https://storybook.js.org/docs/react/writing-docs/autodocs). */
   AUTODOCS = "autodocs",
-  /** Disable global story decorators, applied in `.storybook/preview.tsx`. */
-  DISABLE_GLOBAL_DECORATORS = "disable:globaldecorators",
 }
 
 export default Tags;
