--- src/providers/index.ts
@@ -1,5 +1,0 @@
-export { default as SigilProvider } from "./SigilProvider/SigilProvider";
-export * from "./SigilProvider/SigilProvider";
-
-export { default as ToastProvider } from "./ToastProvider/ToastProvider";
-export * from "./ToastProvider/ToastProvider";
